backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
groq==0.11.0
httpx==0.27.2
google-generativeai==0.8.3
firebase-admin==6.5.0